+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Remove _payload_ from the document list data
 - Improve logs schema
 - Drop a queue and all its related contents
+- Drop a document from the queue
+- [wip] Update queue counters while "playing" a document
+- [pln] Update queue counters while dropping
 
 ## v0.0.6